setup-environment: add new file
This script can be used to initialize the default environment in order to
prepare OE builds. The user will be offered the choice of valid machines, and
some default configuration will be set.

The script also manages the QCOM EULA which is needed on some platforms to
download the proprietary firmware blob.

It can used in different ways:

$ source setup-environment
$ QCOM_EULA=1 source setup-environment
$ MACHINE=ifc6410 source setup-environment
$ source setup-environment my-custom-build

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>